chore: rename function eval_eigenvalues
Renamed function eval_eigenvalues into eval_sub_array_max which is more general and explicative of the real work carried out by the function.
This is done to reuse the function in order to evaluate the maximum value of the norm of the solution variation wrt the previous time step and exploit it in the new recipe for the adaptive time step.
The docstring is updated accordingly.

modified:   utility_functions/transient_solution_functions.py