feat: new funciton interp_at_t_save
Function that performs a linear interpolation in from t_save_left to t_save_right, in order to evaluate values at t_save exploiting array-smart notation. This function is not robust.
This function replaces np.interp1d. Indeed, due to data organization, it is not possible to exploit np.interp1d in array smart notation but a for loop is needed (see exhample below). Since this function si built to address this issue with the data organitazion, it is array smart and it is faster than np.interp1d.

        % Exhample of usage of function np.interp1d.
        % data is a (NN,3) np.array; first column (idx 0) stores values at t1,
        % second column (idx 1) stores values at t2, third column (idx 2)
        % stores interpolated values at tt
        xp = np.array([t1,t2])
        for ii in range(NN):
            fp = np.array([data[ii,0],data[ii,1]])
            data[ii,2] = np.interp(tt,xp,fp)

modified:   utility_functions/auxiliary_functions.py