fix: T_cs evaluated from current_along
In the previous bersion the jop was evaluated as:
jop = (
        np.abs(dict_dummy["op_current"])
        / (self.cross_section["sc"])
        )
where op_current is the current assigned by the input files.
This has been changed to:
jop = (
        np.abs(dict_dummy["current_along"])
        / (self.cross_section["sc"])
        )
where current along is the fraction of current flowing inside the stack
or strand component. For the calculation of T_cs, the current density
was evaluated in this way.

modified:   strand_component.py