feat: new function time_and_event_synchronization
Function that syncronizes the time and the time of the event if their difference in absolute value is smaller than the uncertainty associated to the time step, i.e. |t_{k+1} - t_e| < epsilon.
If it is the case, the event is associated to time t_{k+1} so that t_{k+1} = t_e withouth changing the time step length (since the event is in the neighborhood of t_{k+1}).
When an event is detected the next time step is forced to be the minimum time step defined by the user.

modified:   utility_functions/transient_solution_functions.py