feat: new function force_time_step
Function that forces the time step if the value of the time step
computed with function get_time_step is such that the next event in the
time line will be missed.
This occurs when, with the value for dt computed by function get_time_step,
the time t_k and t_{k+1} are such that
t_k < t_e < t_{k+1}.
with t_k the current time, t_{k+1} the next time and t_e the time at
which an event shoudl occur (like the benning or the ending of a
localized heating).
In this situation t_e would be missed, so a new forced value for dt
is evaluated as dt_f = t_e - t_k. Then the next time steps are
t_{k+1} = t_e
t_{k+2} = t_e + dt_f
The value of dt_f is controlled with tstep_min_lb, see the description for
further details.

modified:   utility_functions/transient_solution_functions.py