feat!: new method __check_event_time_aux_input
Private method that checks if each difference between the time collected in variable l_event can be discretized with at least 10 time steps. If it is not the case, an error is raised. Those values come form the auxiliary input files for heating and/or current. The difference is evaluated as l_event[1:] - l_event[:-1].

Class: Conductor
modified:   conductor.py

BREAKING CHANGES

    Raise an error if time step or minimum time step are non suitably selectet by the user. Before this change if time step or minimum time step were not suitably selectet no error was raised.