chore!: simplify output path
Modified class Simulation to simplify the set of folder in which the output is organized in OpenSc2. In particular not used folders Space_convergence, Time_convergenge and Benchmark will be no longer created by the code.
To achieve this result, several methods of class Simulation are removed and/or modified.

List of removed methods:
    * _space_convergence_paths
    * _time_convergence_paths
    * _benchmark_paths
    * _do_nothing

List of modified methods:
    * simulation_folders_manager
    * _subfolders_paths

Class: Simulation
modified:   simulation.py

BREAKING CHANGES:

The path tree in which the output of OpenSc2 is organized is modified removing not useful folders Space_convergence, Time_convergenge and Benchmark.