fix: ValueError in _compute_density_and_mass_flow_rates
ValueError: No outputs were able to be calculated

The error is raised in method _compute_density_and_mass_flow_rates of class Coolant but it is strictly related to the following waring

MatrixRankWarning: Matrix is exactly singular

generated when function spsolve is called in function electric_steady_state_solution.
The singularity of the matrix is due to some NaN values in the electric stiffness matrix that are related to the electric resistance matrix.
NaN values in electric resistance matrix comes from the fact that method get_electric_resistance of class StackComponent did not return any values as a consequence of a wrong merge conflict management.
This commit reintroducing the return in method get_electric_resistance of class StackComponent fixes this error.

Class: StackComponent
modified:   stack_component.py