fix: update rho_el of stab material for stack object
Updated electrical resistivity of not sc materials (stabilizer) at each electric time step also for instances of class StackComponent.
Changes in methods:
    * operating_conditions_em
    * __eval_gauss_point_em

N.B. This commit should be attributed to A. Mortara who introduced it in a pull request but these commits were not pulled before start merging. This is why they are added at the end of the merging process.

Class: Conductor
modified:   conductor.py