chore: assign current along at t=0 in get_current
Current alogn is assigned in both nodal and Gauss points at t=0, and set equal to op_current for each I0_OP_MODE. The variable current_along represents
the current that actually flows in current carriers, i.e. it is the solution of the electric module of each current carrier component.

N.B. This commit should be attributed to A. Mortara who introduced it in a pull request but these commits were not pulled before start merging. This is why they are added at the end of the merging process.

Class: SolidComponent
modified:   solid_component.py