feat: update method electric_solution_reorganization
Extrapolate the current at the nodes. This is useful for
correctly estimating the current sharing temperature by using the
current actually carried by the current carrier components.
(Prior to these changes, the input current was used, but this
could lead to over- or underestimation of the current sharing
temperature.)

Class: Conductor
modified:   conductor.py