fix: correct jop evaluation in method eval_tcs
Evaluate current density exploiting the current that is actually carried by the SC current carrier (Stack or StrandMixed) in order to correctly evaluate the current sharing temperature for each Stack and or StrandMixed.

Since the current sharing temperture is evaluate both in nodal and Gauss points, the previous commit are added in order to have key current_alogn also available in attribute dict_node_pt.

N.B. This commit should be attributed to A. Mortara who introduced it in a pull request but these commits were not pulled before start merging. This is why they are added at the end of the merging process.

Class: StrandComponent
modified:   strand_component.py