chore: add tolerance in solve_current_divider
Added tolerance in method solve_current_divider to avoid inaccuracies on the divider that could lead to voltage differences between sc and stab that are not expected in the parallel of electric resistances when the critical current is very small.

N.B. This commit should be attributed to A. Mortara who introduced it in a pull request but these commits were not pulled before start merging. This is why they are added at the end of the merging process.

Class: StackComponent
modified:   stack_component.py